cleanup legacy package a little
move types where their functions are
